requirement.txt
asgiref
async-generator
attrs
backports.zoneinfo
blinker
certifi
cffi
charset-normalizer
click
colorama
cryptography
cycler
distlib
exceptiongroup
filelock
Flask
Flask-Cors
Flask-SQLAlchemy
fonttools
gevent
greenlet
h11
idna
importlib-metadata
itsdangerous
jieba
Jinja2
joblib
kiwisolver
lxml
MarkupSafe
matplotlib
mysqlclient
numpy
outcome
packaging
pandas
Pillow
platformdirs
pycparser
PyMySQL
pyOpenSSL
pyparsing
PySocks
python-dateutil
pytz
requests
scikit-learn
scipy
selenium
six
sniffio
sortedcontainers
SQLAlchemy
sqlparse
threadpoolctl
trio
trio-websocket
typing_extensions
tzdata
urllib3
urllib3-secure-extra
virtualenv
virtualenvwrapper-win
Werkzeug
wordcloud
wsproto
zipp
zope.event
zope.interface